+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/model/FapCategory.kt
@@ -7,6 +7,8 @@ import kotlinx.serialization.Serializable
 @Parcelize
 @Serializable
 data class FapCategory(
+    val id: String,
     val name: String,
-    val picUrl: String
+    val picUrl: String,
+    val applicationCount: Int
 ) : Parcelable