+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/model/FapItemVersion.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.faphub.dao.api.model
+
+import com.flipperdevices.core.data.SemVer
+
+data class FapItemVersion(
+    val id: String,
+    val version: SemVer
+)