+++ components/faphub/installation/queue/impl/src/main/java/com/flipperdevices/faphub/installation/queue/impl/api/FapInstallationQueueApiImpl.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.faphub.installation.queue.impl.api
+
+import com.flipperdevices.faphub.installation.queue.api.FapInstallationQueueApi
+
+class FapInstallationQueueApiImpl : FapInstallationQueueApi