+++ components/faphub/installation/queue/impl/src/main/java/com/flipperdevices/faphub/installation/queue/impl/executor/FapActionExecutor.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.faphub.installation.queue.impl.executor
+
+class FapActionExecutor