+++ components/faphub/installation/queue/impl/src/main/java/com/flipperdevices/faphub/installation/queue/impl/executor/actions/BundleDownloader.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.faphub.installation.queue.impl.executor.actions
+
+class BundleDownloader