+++ components/faphub/installation/queue/impl/src/main/java/com/flipperdevices/faphub/installation/queue/impl/model/FapActionRequest.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.faphub.installation.queue.impl.model
+
+sealed class FapActionRequest {
+    data class Install(
+        val applicationId: String,
+        val toVersionId: String
+    ) : FapActionRequest()
+}