+++ components/faphub/target/api/src/main/java/com/flipperdevices/faphub/target/api/FlipperTargetProviderApi.kt
@@ -1,0 +1,9 @@
+package com.flipperdevices.faphub.target.api
+
+import com.flipperdevices.faphub.target.model.FlipperTarget
+import kotlinx.coroutines.flow.StateFlow
+
+interface FlipperTargetProviderApi {
+    fun getFlipperTarget(): StateFlow<FlipperTarget>
+    suspend fun getFlipperTargetSync(): Result<FlipperTarget.Received>
+}